fix: moved 'syncNodeGraph` from `NodeConnectionManager` to `NodeManager`
